+++ guillotina/tests/cache/test_cache_memory.py
@@ -93,7 +93,7 @@ async def test_cache_object(guillotina_main):
     cache = BasicCache(txn)
     txn._cache = cache
     ob = create_content()
-    storage.store(ob)
+    storage.store(None, None, None, ob, txn)
     loaded = await txn.get(ob.__uuid__)
     assert id(loaded) != id(ob)
     assert loaded.__uuid__ == ob.__uuid__
@@ -115,8 +115,8 @@ async def test_cache_object_from_child(guillotina_main):
     ob = create_content()
     parent = create_content()
     ob.__parent__ = parent
-    storage.store(parent)
-    storage.store(ob)
+    storage.store(None, None, None, parent, txn)
+    storage.store(None, None, None, ob, txn)
 
     loaded = await txn.get_child(parent, ob.id)
     assert cache._hits == 0
@@ -136,7 +136,7 @@ async def test_do_not_cache_large_object(guillotina_main):
     txn._cache = cache
     ob = create_content()
     ob.foobar = "X" * cache.max_cache_record_size  # push size above cache threshold
-    storage.store(ob)
+    storage.store(None, None, None, ob, txn)
     loaded = await txn.get(ob.__uuid__)
     assert id(loaded) != id(ob)
     assert loaded.__uuid__ == ob.__uuid__