+++ CHANGELOG.rst
@@ -8,8 +8,47 @@ CHANGELOG
   [dmanchon,masipcat,vangheem]
 
 
-5.0.4 (unreleased)
+5.0.9 (unreleased)
+
+- Nothing changed yet.
+
+
+5.0.8 (2019-09-05)
 ------------------
+
+- pypi package desc fix
+
+
+5.0.7 (2019-09-05)
+------------------
+
+- Explicitly reset task vars on every request
+  [vangheem]
+
+- Fix futures execute error when no futures are defined for type
+  [vangheem]
+
+
+5.0.6 (2019-09-04)
+------------------
+
+- Fix `execute.clear_futures()`
+  [vangheem]
+
+- Adding Helm Charts
+  [karannaoh]
+
+5.0.4 (2019-09-04)
+------------------
+
+- Upgrade mypy
+  [vangheem]
+
+- Fix not setting cache values for updated object when push is not enabled
+  [vangheem]
+
+- Fix conflict error handling with registry objects
+  [vangheem]
 
 - Sorted imports in all files and added `isort` in .travis to keep the format
   [masipcat]
@@ -88,6 +127,8 @@ CHANGELOG
 - Fix cache key for SQLStatements cache. This was causing vacuuming on multi-db environments
   to not work since the vacuuming object was shared between dbs on guillotina_dynamictablestorage.
   [vangheem]
+
+- Refractor and bug fix in validation of parameter
 
 - Implement more optimized way to vacuum objects which dramatically improves handling
   of deleting very large object trees